- move analytics to general components folder
