+++ components/analytics/Registration.jsx
