+++ pages/bootcamps/current/[slug].jsx
@@ -3,8 +3,8 @@ import React from "react";
 import Navbar from "../../../components/layouts/Navbar";
 import Footer from "../../../components/layouts/Footer";
 import bootcampsData from "../../api/bootcamps.json";
-import Registration from "../../../components/bootcamps/analytics/Registration";
-import Tracks from "../../../components/bootcamps/analytics/Tracks";
+import Registration from "../../../components/analytics/Registration";
+import Tracks from "../../../components/analytics/Tracks";
 import GoalsDetails from "../../../components/bootcamps/Goals";
 import HeroSection from "../../../components/bootcamps/HeroSection";
 import About from "../../../components/bootcamps/About";
