+++ pages/bootcamps/upcoming/[slug].jsx
@@ -2,8 +2,8 @@ import { useRouter } from "next/router";
 import React from "react";
 import Navbar from "../../../components/layouts/Navbar";
 import bootcampsData from "../../api/bootcamps.json";
-import Registration from "../../../components/bootcamps/analytics/Registration";
-import Tracks from "../../../components/bootcamps/analytics/Tracks";
+import Registration from "../../../components/analytics/Registration";
+import Tracks from "../../../components/analytics/Tracks";
 import GoalsDetails from "../../../components/bootcamps/Goals";
 import HeroSection from "../../../components/bootcamps/HeroSection";
 import About from "../../../components/bootcamps/About";
